feat(iOS): Support modifying the AFI on iOS
Adds a new method, writeAfi, which allows the writing of a tag's AFI, provided the tag supports it.
This method requires and existing session, so using `scanTag` with `keepSessionOpen` set to `true` is recommended.

Also ports a ProcessNFCV method written by Brent Vidler, important for getting the content of ISO15693 tags but not
necessarily related to AFI setting.